fix test

+++ config/mail.php
@@ -41,7 +41,7 @@ return [
             'transport' => 'smtp',
             'url' => env('MAIL_URL'),
             'host' => env('MAIL_HOST', 'localhost'),
-            'port' => env('MAIL_PORT', 1025),
+            'port' => (int) env('MAIL_PORT', 1025),
             'encryption' => env('MAIL_ENCRYPTION', 'tls'),
             'username' => env('MAIL_USERNAME'),
             'password' => env('MAIL_PASSWORD'),
